fix: load correct list history
